pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "o11y-bench"
version = "0.1.0"
description = "Open benchmark for evaluating LLM agents on observability and SRE tasks"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "Markdown>=3.10",
    "anthropic>=0.40",
    "harbor==0.3.0",
    "pydantic>=2.0",
    "promql-parser>=0.7.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
    "mypy>=1.10",
    "types-Markdown>=3.8",
    "types-pyyaml>=6.0",
]

[tool.hatch.build.targets.wheel]
only-include = ["agents", "grading", "o11y_bench", "reporting", "scripts"]

[tool.pytest.ini_options]
testpaths = ["tests", "docker/python/tests"]

[tool.ruff]
target-version = "py314"
line-length = 100
extend-exclude = ["tasks"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "RUF"]
ignore = [
    "ANN401",  # dynamically typed expressions (Any) are disallowed — too strict for tool schemas
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN"]  # don't require annotations in tests
"docker/python/tests/*" = ["ANN"]  # don't require annotations in Docker-side tests
"agents/o11y_agent.py" = ["ANN"]  # Harbor agent (imports resolved at runtime)
"reporting/**" = ["ANN", "E501"]
"o11y_bench/**" = ["ANN", "E501"]
"scripts/**" = ["ANN", "E501"]
"grading/**" = ["ANN", "E501"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
files = [
    "agents/o11y_agent.py",
    "agents/agent_runner.py",
    "grading/checks.py",
    "grading/dashboard_queries.py",
    "grading/dashboard_snapshot.py",
    "grading/dashboard_state.py",
    "grading/facts.py",
    "grading/judge.py",
    "grading/models.py",
    "grading/scoring.py",
    "grading/transcript_parser.py",
    "reporting/compare_report.py",
    "reporting/categories.py",
    "reporting/report.py",
    "reporting/report_data.py",
    "reporting/report_paths.py",
    "reporting/run_report.py",
    "reporting/summary.py",
    "o11y_bench/cli.py",
    "o11y_bench/config.py",
    "o11y_bench/harbor.py",
    "o11y_bench/resume.py",
    "o11y_bench/run.py",
    "scripts/sync_tasks.py",
    "grading/verifier.py",
]

[[tool.mypy.overrides]]
module = [
    "promql_parser",
    "harbor.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["agents.o11y_agent"]
disable_error_code = ["misc"]
